backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.9.1
requests==2.32.3
aksharamukha==2.3
python-dotenv==1.0.1
Pillow==10.4.0
pytesseract==0.3.13
aiofiles==23.2.1
python-multipart==0.0.9
loguru==0.7.2
httpx==0.28.1